requirements.txt
streamlit==1.29.0
ccxt==4.1.93
pymongo==4.6.1
python-dotenv==1.0.0
schedule==1.2.1
pandas==2.1.3
requests==2.31.0
matplotlib==3.8.2